requirements.txt
pdfplumber>=0.9.0
python-pptx>=0.6.20
openpyxl>=3.0.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.20.0
regex>=2022.0.0
transformers>=4.41.0
torch>=2.2.0
opencv-python>=4.9.0.80
python-docx>=1.1.0
pytesseract>=0.3.10
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
